requirements.txt
alembic==1.18.0
blinker==1.9.0
cfgv==3.5.0
click==8.3.1
distlib==0.4.0
filelock==3.20.3
Flask==3.1.2
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
identify==2.6.16
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.0
nodeenv==1.10.0
platformdirs==4.5.1
pre_commit==4.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
SQLAlchemy==2.0.45
typing_extensions==4.15.0
virtualenv==20.36.1
Werkzeug==3.1.5
